requirements.txt
pyyaml==6.0.1
Playwright==1.42.0
